requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.118.0
google-auth==2.41.1
google-genai==1.41.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
